move main to main.py, load classifiers from yaml
