+++ requirements.txt
@@ -21,3 +21,4 @@ scipy==1.3.1
 six==1.12.0
 wcwidth==0.1.7
 zipp==0.5.2
+pyyaml=5.1.2
